saving SCORM data and managing attempts works!
When a student accesses the LTI, we check if they have any attempts at the resource. If yes, list the attempts. If no, create an attempt and run it.
From the list of attempts, you can resume (or review) one, or start a new attempt.

Overwriting an exam shouldn't break existing attempts - they'll still use the old package.

Need to add security checks on the WebSocket consumer so people don't send data to other people's attempts, and to make the websocket handling in the SCORM API more robust.