deal with broken attempts
we've seen a few instances of attempts becoming unusable because only a couple of SCORM elements are saved. In particular, cmi.suspend_data is missing.
The most straightforward explanation is that the user closes the window before everything saves, but I think it's more likely to be an error on the server side.
For now, this commit adds code to check if the attempt being resumed is broken, and if so silently create a new one, marking the old one as broken.

I should add some error-catching and reporting to the SCORM API websocket, so if something goes wrong the user is told.