add web-based setup and consumer management
When you load the index for the first time, you're prompted to create a
superuser, and then a first LTI consumer.
After that, there's a link on the index to a "manage consumers" page,
where you can get the key/secret, and add/delete consumers.
There's also a link to the django admin, for more fiddly tasks.