filter available exams, and record when exams are imported from the editor
On the create exam page, there's a search field to filter the list of
exams shown.

When an exam is imported from the editor, its URL on the REST API and
the URL used to get the scorm package are saved, so it can be updated
later on.

There's now an option on the replace exam package to update from the
editor, when an exam was imported directly from the editor.