Saving is much more robust
First of all, this replaces ReconnectingWebSocket with RobustWebSocket, a different implementation of the same idea that claims to be a bit more... robust.

The point of this commit is to add in a load of fallbacks to cope with connectivity problems when trying to save SCORM data.
The first problem is that WebSockets don't guarantee delivery of messages, so the server sends a receipt when it has saved data.
The client keeps a dict of data which has been sent but for which it hasn't had a receipt yet, so it can be resent if needed.
Unconfirmed data is resent whenever the websocket reconnects.

If the websocket doesn't work for a long time, there's now an AJAX fallback which runs every 5 seconds when there's no websocket connection.
The AJAX fallback behaves the same way as the websockets, saving data until it gets a confirmation it's been saved.
If the websocket reconnects, it can quickly resend any data the AJAX fallback didn't get a receipt for.

Finally, we try to save all this data to localStorage, so if everything goes wrong, the client can still resume where it left off later on.
The SCORM API merges the elements sent by the server with the data saved in localStorage. Each element has a timestamp, so we can work out the latest value of each element.

It's possible that none of this code ever runs, because of a syntax error or something unforeseen with an unusual browser, so the whole thing is wrapped in a try/catch
which will tell the student there's a problem and redirect to the attempt summary page.
This should avoid the case where nothing ever works, but the SCORM package still loads and looks like it's working, but never interacts with the SCORM API because it hasn't been inserted into the page.