add counter to disambiguate elements with the same timestamp
when the same scorm element is set repeatedly in quick succession (for
example, cmi.score.raw is set several times when submitting all parts in
a question), several elements with the same timestamp can be created.

This adds a 'counter' field to scorm elements, which increases in api.js each time
an element is set. This doesn't need to be consistent between sessions -
it only needs to disambiguate elements sent from the same client at the
same time.