big red warning when exam ended and still unsaved
The warning message becomes bigger and permanently displayed at the top
of the window if the exam is ended and there is still unsaved data.
This should prevent students leaving the exam without noticing that data
is unsaved - if they've finished, we can trade the inconvenience of a
stonking warning sign getting in your way for making sure they haven't
missed it.

I also caught a case where the socket claims it's open but it definitely
isn't - I'd turned networking off. The socket_is_open method now returns
false if there's any sent data that hasn't been acknowledged - if this
remains the case for long enough, then we have to assume the socket is
broken.