Keep track of lots more info from LTI
The LTI post sends a lot of information about the course the activity is
being launched from, which we can use to organise resources in the admin
view.

Now every resource belongs to a `context`, which maps onto Moodle or BB
courses.

In the admin view, each consumer has a management page, where you can
see all the resources belonging to that consumer, organised by context.
Clicking on a resource takes you to its instructor view.