disambiguate resource on instance_guid, not context
We can't rely on a resource staying in the same context on the consumer,
so instead use the consumer's tool_consumer_instance_guid.

(tool_consumer_instance_guid, resource_link_id) should be globally
unique.

Ideally I'd use the consumer key as well, but django_auth_lti doesn't
save that in request.LTI, and relying on it would require invalidating
everyone's sessions, which I don't want to do in the middle of term!