use get_or_create when making ScormElements
I noticed that when a request to save scorm elements fails, the client
has no way of knowing if any of the elements in the batch were saved, so
sends them all again.
The server should avoid creating copies of elements. I thought
get_or_create would be too expensive, but it looks like it's OK.