trying to make it clearer what you need
Most of the people looking at the LTI provider probably aren't in a
position to set it up themselves. This adds a section "What do I need in
order to use this?" which provides enough information for someone to
decide if they're capable, or need to pass it on to their IT support.