catch DB operational error in SCORM fallback
We recently encountered some OperationalErrors caused by trying to save
Unicode characters to a MySQL database with the wrong encoding (utf8
instead of utf8mb4)

This patch catches those errors in the SCORM fallback, and fixes the
error catching in the JS so they cause the warning to appear.

This also adds SUPPORT_NAME and SUPPORT_URL settings, which are shown
in the warning box as a link to click on if the warning stays visible
for a long time.