log db.OperationalError caused by invalid value when saving a scorm element, but allow other elements to be saved
Our MySQL database is set up wrong, so fails to save 4-byte unicode characters, throwing OperationalError with code 1366.
This commit catches that error, logs it, and sends back a list of unsaved elements to the client. At the moment, the client does nothing with that information - we assume the data will never be saved.