some changesto the remark part view
There's a "back to attempts listing" link at the top of all attempt
views now, fixing #37.

You can press enter in the input box to remark a part to save. The box
goes red when the score is changed and hasn't been saved.

This is still pretty opaque.
Maybe this should just be a long page with a Save button at the bottom.