heatmap of question scores in attempt list
helps with #3

To achieve this, info on question scores per attempt is now saved in
AttemptQuestionScore objects. I added that model a while ago but didn't
use it because I couldn't work out how to keep it up to date.

Now, these objects are set whenever a cmi.objectives.N.score.X SCORM
element is saved, or when a RemarkPart or DiscountPart object is saved
or deleted.

They're created on-demand for old attempts, so I don't have to run an
enormous DB migration.