store end_time for attempts so extra elements can't be saved after end
fixes #24

Attempt has a field `end_time` which is set when `cmi.completion_status`
is set to "completed".

When an attempt is completed, a websocket message is sent to all
connected clients, telling them to end. An alert is shown to the user,
and then the page is reloaded (so the exam reopens in review mode).

There's still the problem of attempts getting out of sync if open in two
windows at once. I think there should be a 'lock' on attempts, and on
opening a new window any other windows should quit.