kill a SCORM client if another one is opened elsewhere
SCORM clients identify themselves with a unique ID when they connect,
generated as timestamp+Math.random().
If the attempt is not opened in review mode, the server sends a message to all connected clients for the attempt with
that ID. Any clients with different IDs disconnect and navigate to the
attempts listing page.

This also changes the external_kill logic so that attempts open in review
mode do not get killed.

improves on #24