still try the fallback if the socket is open
In the last commit I made send_ajax return if the socket is open, and
now I remember that's not right: if there's nothing to send, it'll quit
anyway.
If there's data that has been sitting for more than the ajax_period, we
should try AJAX. send_ajax now doesn't try to start new requests until
the previous one has either completed or failed, so we won't end up
hammering the server with the same data.