add time periods per consumer
The consumer management page now has a link to a form to define time
periods for the consumer.

Once time periods have been defined, contexts are grouped by the time
period they belong to: the earliest date of creation of a resource in
each context is used, and contexts go into the period with the latest
end date.

If contexts fall into a gap between periods, their shown separately.

fixes #53