don't abort reporting process on error
When reporting all scores for a resource, save errors produced for
individual users and report them in bulk at the end.

This way, we make the best effort to report everyone's scores, ignoring
errors corresponding only to some students.

We got "invalid role" errors from Blackboard when a student is no longer
on the course (or enrolled at the uni?)