when resuming a completed attempt with unsaved completion status, redirect back to attempts screen
If a student completes an exam and closes the browser before the
completion status is saved to the server, they'll still be able to
resume it. If they resume before review mode is available, they
shouldn't be allowed into the exam.
The SCORM API, on merging the saved data with that in localstorage,
saves the data with an AJAX request then redirects back to the attempts
listing. The student is shown a message explaining what happened.