stuff about opening in a new window
Cross-site security policies can prevent iframes from writing cookies,
which scuppers the LTI tool's session management.

If the LTI activity launches in a new window instead, the browser will
allow it to write cookies, so everything works.

This commit adds a hint to the "not authorized" page about opening in a
new window, and some lines to the Moodle and Canvas setup instructions.

Annoyingly, while Moodle lets you set up LTI tools to launch in a new
window by default, in Canvas it looks like you have to set it for each
individual activity.