send email receipts when attempts completed
When the student completes an attempt, they can be sent an email receipt
confirming the completion.

The receipt contains start and end times, and the raw score if the
student's allowed to see it.

The receipt includes a cryptographically-signed code which the
instructor can use to validate the infromation in the student's receipt.

Receipts are only sent if the setting EMAIL_COMPLETION_RECEIPTS is True

Fixes #67