add "available from" and "available until" resource settings
fixes #29

If "available from" is set, students may not start attempts until that
time.

If "available until" is set, students may not start attempts after that
time, and any attempts in progress at that time will be automatically
ended - the exam will disappear, and a message saying "This attempt has
closed" will be shown. There is also a message about data saving -
either that it's in progress, or it's done and you can quit.