more receipt fiddling
There was *still* a race condition when sending the email receipt: the
end_time was set by a signal looking at ScormElements, but the
'complete' flag was checked during the AJAX response, which happens
before signals.

Now, the Attempt model has an 'all_data_received' flag which is set
during the AJAX response. There's a separate signal called when attempts
or scorm elements are saved, which checks the conditions for sending an
email:

* all data received
* end time set
* completion status is 'completed'
* no email sent yet

If those are all true, an email is sent.

This commit also adds a receipt code to the response of the final AJAX
save call, which the browser displays, so students can confirm their
data even if they don't get the email.