shift jobs to workers; fix email race condition
There are now consumers for the tasks of reporting a user's score
immediately on attempt completion, and emailing a completion receipt.

The signal which decides when to send a completion receipt now fetches
the attempt object from the database again, to avoid race conditions
where the object is updated in quick succession. The methods that do
this also use the `update_fields` keyword on `.save` so that they only
save the changed fields.

This might fix #86