set a timeout on all requests; handle all request errors when reporting outcomes
fixes #97

This commit adds a new setting REQUEST_TIMEOUT to settings.py, which is
the number of seconds to wait for a request to time out. The default is
60.

The "report outcomes" process now handles timeouts, and any other
exception coming from requests. If you have tried to cancel a reporting
process and it's still running, you're shown a warning message that you
should wait for it to finish before trying again.
There are circumstances in which you might want to try again before the
last process is marked as complete - restarting the server, you've fixed
the underlying problem, etc.