automatic testing of packages, and remarking
Working towards #110: when an exam package is uploaded or replaced, test
it using the Numbas testing framework. At the moment the upload form is
invalid if this test fails; there should be an option to ignore this.

There's a new management command `test_exam` which can be used to test an
exam package from the command line.

Working towards #98: there's a management command `remark` which will
remark all attempts at a resource. It will open each attempt in review
mode using the Numbas testing framework, and report any SCORM elements
that change.
I still need to sort out how attempts using old exams will work. There
could be an option to use the resource's current exam package, or the
package linked to the attempt.