Starting work on web-based remark view
It occurred to me that an instrructor would want to start the remarking
process through the web interface, and if we've got a browser we might
as well use that to do the marking. It avoids putting load on the
server, too.

This commit adds a page to remark all the attempts at a resource against
the current exam package. It can do the remarking, but I haven't written
the code to save changed data back yet.