improving the attempt timeline view
Elements within a second of each other are grouped together, instead of
requiring exactly the same time.

The score column is only shown when the score changes, to more easily
identify times when the score changes.

The calls to get stuff from the datamodel from a certain time now just
run through looking for a particular element, instead of building the
whole data model.