set the real exam start time from attempt data
Until now, the Attempt.start_time field was set when the attempt record
is created, but the student might not click "Start" in the Numbas
interface until much later.

This adds a signal listening for the cmi.suspend_data SCORM element, and
updates start_time using the 'start' key in that if it's different.