store cmi.suspend_data as diffs
The cmi.suspend_data SCORM element is very big, and rarely changes. To
save on storage, this commit changes the ScormElement model so that its
value can be stored as a diff relative to another ScormElement object.

When data is received from the student, it is saved in full. A periodic
huey task looks for attempts with more than one cmi.suspend_data
element, and rewrites the older ones as diffs relative to newer ones.

The most recent value is always stored in full, so it can be retrieved
easily.

When the full SCORM history is displayed, in either the timeline or
SCORM data views, any diffed elements are resolved to their full values.

fixes #130