controls to manage versions of exam packages
This adds a 'resource' field to exams, so old versions can still be
associated with the resource they were uploaded to.

The "replace exam package" page now lists all of the exams associated
with a resource, and how many attempts use each of them. You can change
the active exam package, and set existing attempts to use the active
version.

Still to do: run an attempt using a particular version, set the version
used by a particular attempt.

see #113