check that the session cookie has been set on launch
fixes #118

When the tool is launched in an cross-domain iframe, some browsers
refuse to set or send cookies.
This adds an intermediate step to lti_entry which checks that the
session cookie has been set. If it hasn't, it shows a link to open in a
new tab a URL which will try again to set the session cookie.

Once the cookie has been set, it performs the lti_entry logic.