a different tack on diffing suspend data
Attempt records now have a 'diffed' boolean field, recording whether
they have any cmi.suspend_data elements that might need to be diffed.
It's set to 'false' when elements with that key are saved.

This makes it much quicker to find attempts that need to be diffed, so
the task should be able to get to work much quicker.

see #133