exam duration extensions
This relies on some changes in the Numbas runtime: the API does a
postMessage to the runtime's frame, to tell it to check the SCORM model
for new data.

see #144

Access changes can now specify an extension to the duration of the exam,
in either percent or minutes.

The logic around changing availability dates is now a bit simpler: it
just finds the earliest available_from and the latest available_until,
instead of returning a list of intervals. I don't think anyone will
really want to allow access in disjoint intervals.

I'm still not sure what to do when available_from > available_until. I
can't think of any circumstances in which you'd want to change the dates
for only some students.