use the last created access change.
Using earliest available_from and latest available_until was a bad
idea, immediately: we had a student for whom we wanted to give a later
window of access to complete a test that they missed.

On setting up an access change, I saw that it would in fact be available
fom the original start time until the new end time.

This changes the logic so that when it runs through the list of access
changes, any changed value overwrites the current value, for both
availability times and duration extensions.