when running an attempt, in the first instance use the LTIUSerData object to check permission
The RunAttemptView was using the session's saved LTI data to check if
the user is an instructor, but if the session isn't working properly
that doesn't work. We didn't work out why the session was messed up, but
it was repeatable.

The LTIUserData object for the user and resource is the first place to
check if the user has instructor privileges for the resource.