deal with errors and delays in getting an attempt receipt
fixes #181

Once all data has been acknowledged because it was sent over websocket,
but still waiting for the receipt to be sent over AJAX, the "please
wait" message is shown, in red.

If there's an error with the final AJAX call that gets the receipt, a
message "please ask your instructor to check your data is saved
correctly" is shown.