use height to show correctness on attempt score info
In the resource attempts listing, there's a visualisation of the
student's score for each question.

Until now, only colour was used to show the different grades. This
commit changes the height of each bar, so that completely correct
questions are full height and completely incorrect are just a couple of
pixels high.