discount parts is a simple form; do score recalculation as a task
Like the individual attempt remarking page, the resource-wide part
discount page is now a simple form.

There are no longer signals hooked up to RemarkPart and DiscountPart
objects - the forms trigger a task to recalculate scores when they're
submitted.

This should make the views a lot more responsive. For resources with a
lot of attempts, there could be a problem if another change is
made before the previous one finishes recalculating scores.