trying to get tasks to work properly!
The signals triggered by saving a SCORM element queued a task which
would try to fetch the latest version of the SCORM element from the DB.
Because the signal is triggered inside a transaction, if the task
happens before the transaction is finished then the SCORM element won't
be found.

This changes the call to send the element itself, assuming that if
another element is saved before the task runs then the newer element
will have queued another run of the task anyway.