trying to avoid out-of-order ScormElement diffs
see #194

Something happens to cause numbas_lti.models.diff_scormelements to diff
a newer element against an older one, messing up the chain of
dependency.
I have no idea how that can happen, but for now, this commit adds a line
to just not make the diff when it looks like the order is wrong.