when saving SCORM data, update score info when total changed and when finalising attempt
should fix #193

In numbas_lti.save_scorm_data, the task to update the score info for an
attempt was only called when a question score changes. The total raw
score is also used to save the cached scaled_score, so the task should
run when the total changes too.

To be extra sure, the task also runs when an attempt is finalised.