show the available until date during an attempt
fixes #202

If the resource has an "available until" date set, a banner saying "This
attempt will automatically close at <X>, if not already closed."

This commit also adds a "try again" button under the message "there was
an error closing this attempt". IF the final AJAX call fails, that
message is shown. Clicking the button makes it go again. This should
reduce the number of emails we get asking if data was saved correctly.