attempt timeline: links to view the attempt at that time
The time next to each group of elements in the timeline view is now a
link, which will load the attempt with all the data up to that point.

I've added this because a student's attempt has broken, and I want to
step through their timeline to see exactly where.