add "require lockdown app" option to resources
This adds an option to resources to require the Numbas lockdown app.

When a student opens a resource which requires the app, they will be
shown a page with a link to a numbas:// URL. The URL contains the LTI
provider's host, and an encrypted settings object.

The settings object is encrypted with a password, which the user has to
type in, and a salt, which is shared between the LTI provider and the
app. (It should be the same globally, since everyone will install the
same version of the app)

There is one password for the whole LTI provider.
This might change if needed, but my thinking is that students shouldn't
have to look up a new password for each assessment.
If a password is needed to unlock a timed or invigilated assessment, it
should be set on the Numbas exam, so that students can get to the exam
front page before the test is due to start.

The encrypted settings object contains the URL to launch, and the
session key and resource link id from the unlocked browser. This reuses
the logic from the cookie check, letting the student pick up their
session data in the lockdown app.

Once the lockdown app has picked up the session, it sends an
Authorization header with every request to the LTI provider, which is
used to decide if the client is using the lockdown app or not.
At the moment, the "show attempts" and "run attempt" views require
this header.