avoid creating duplicate LTIContext records
NumbasLTIResourceMiddleware was doing two separate queries to get an
existing LTIContext object, or create one if it doesn't exist.
This can lead to multiple records being created by separate threads.

This commit changes that code to use the `get_or_create` method in one
atomic transaction, and adds a `unique_together` constraint on
LTIContext so a duplicate can't be created.