Add support for Safe Exam Browser
This adds the ability to require that activities launch in Safe Exam
Browser, as an alternative to the Numbas lockdown app.

In the admin dashboard there are views to manage SEB settings files. The
form shows you the launch URL to use, and then takes a .seb settings
file as well as the config key hash.

Instructors can pick from the available SEB settings files in the
resource's settings form.

When the student launches a resource requiring SEB, they're redirected
to a seb:// URL which will load the settings file in the SEB client,
redirecting to a launch page.

The server checks that the client is using SEB using the
X-Safeexambrowser-Configkeyhash header.

This header isn't sent by the Mac or iOS apps, because of a restriction
in the WKWebView API. SEB recommends a JavaScript method of
authentication, which I haven't implemented. We'd need to think about
how it should work: you don't want to send anything that the student
shouldn't see before authenticating them.

There's a new field to set a resource-specific password when using the
Numbas lockdown app.

There's a setting to show the password to the student on the lockdown
app launch page.