use a global flag to prevent loading cached pre-submit results
Just emptying the pre-submit cache after loading was not enough: parts
that have already been answered have submit() run as part of the loading
process, so the cache already needs to be empty.

This commit changes it so that a global flag
`Numbas.load_pre_submit_cache = false` is set, and the part loading
method obeys that.

Finally, in the re-marking controller, all parts are checked for
`waiting_for_pre_submit` whether or not "use unsubmitted answers" is
ticked, since answered parts with pre-submit tasks will also be waiting.

see #265