Starting to work on adding LTI 1.3 support
My plan is to abstract the link between a resource and an LTI link.

The LTI 1.1 stuff has been moved to prefixed LTI_11_* classes and
methods.

The django_auth_lti method of making a User record whose username
matches the username given by the LTI launch always caused problems, so
I'm taking the opportunity to change. User records will be given a UUID
as a username to satisfy the requirement for a username, and a linking
table is used to relate LTI launch data with the User model. I looked
into using a custom User model but it's really hard to switch to a
different model with an existing database.

Currently:

Working on creating a resource for an LTI 1.3 resource link.

LTI 1.1 launches don't work.

Got to the point where CreateExamView needs to get a resource, which can be found by getting a ResourceLink matching the LTI message_launch data.