The SCORM API sends back times with timezone info
fixes #296

I've been through api.js and checked that all time operations use
luxon objects instead of numbers of milliseconds.

When the API reports SCORM elements back to the server, it represents
the time as an ISO format string with timezone information. This avoids
ambiguous times due to daylight savings.