Load SCORM data through fetch instead of embedding it in the page
fixes #266

The run_attempt view doesn't embed the SCORM data model in the HTML any
more. Instead, the client makes a fetch request to a new `/run_attempt/<pk>/scorm_cmi`
view, which returns the current SCORM data.

This ensures that even if the page has reloaded from the browser's
cache, it will only start with the latest data.
If the data can't be loaded for some reason, such as no network
connection, the student is shown an error message and a button to reload
the page.

This commit also makes the run attempt and scorm cmi views throw a
PermissionDenied error if the attempt would be opened in review mode,
but review is not allowed yet, and the user is not an instructor.
Previously, a student could open an attempt in review mode by reloading
the page, even when it's not allowed yet.

An AttemptLaunch record is created each time the SCORM data model is
requested. This should give more detail about when a student reloads an
attempt by browser navigation instead of by clicking links.