The LTI 1.3 teacher launch can create a resource and link an exam to it
The LTI 1.3 launch id is passed through the views from teacher launch,
to create a resource, create an exam linked to it, and then view the
settings page.

The lti_13_launch_id parameter needs to be added to all links. In
django_auth_lti, they achieved this by patching Django's `reverse`
function, to add an identifier for the LTI 1.1 launch ID automatically,
when it's available.

I'd rather explicitly keep track of which views need LTI launch
information.
I'm not completely sure about how to do this conveniently. I've added a
template tag `url_with_lti`, which gets the LTI message launch object
from the template context, if possible, and a method `reverse_with_lti`
to LTI_13_Mixin. Ideally, it'd be possible to only have one of these,
but I can't see how.