creating a new resource from deep link works
You're directed to upload an exam, and then the deep link flow ends.
The LTI_13_ResourceLink can't be created at this point, because the link
has no ID yet.

When you launch the link, the LTI_13_ResourceLink is automatically created if needed, and updated with the title passed from the launch.