starting work on using NRPS for access changes
The Access Change edit views fetch the list of members from the Names
and Roles Provisioning Service.
The form shows a table of all members with student roles.
I'm starting to add a search box - it needs to be hooked up to filter or
highlight the selected student.

At the moment, the list of members is fetched while rendering the form.
If this takes a long time, then the request might time out. Should we
cache the list of members? How often should it be updated? Maybe as a
task, each time a Teacher launch happens?