fix SEB launch check
This also adds an error message for when a launch request seems to come
from SEB, but with no session parameters. This happens when you launch
SEB using a previously-set config, but not from the link which includes
the session key and resource link ID in query params, or if the SEB
settings file doesn't have "allow query params" turned on.