LTI 1.3 launch: if no LTI 1.3 resource link found, look for a LTI 1.1 link
Blackboard upgrades LTI 1.1 links to LTI 1.3 links by making a launch
with the same context_id and resource_link_id.

This commit changes the lti_13.ResourceLaunchView to look for a matching
LTI_11_ResourceLink if no LTI_13_ResourceLink is found.

The context_id and resource_link_id are only supposed to be unique to
the consumer, so there could be a clash if different consumers use the
same IDs. I think we'd need to add a model linking an LTI 1.1 consumer
with an LTI 1.3 consumer to avoid this ambiguity.