working on documentation
I've been going through the documentation, adding information for LTI
1.3 where necessary, and updating screenshots.

The documentation about registering LTI 1.3 connections is incomplete:
we need to completely record the steps necessary for registering with
Canvas and Blackboard, and the registration page needs to have forms for
saving the deployment IDs. (or there needs to be a form on the consumer
management view).