Tokens for LTI 1.3 dynamic registration
This adds a system of tokens for registering LTI 1.3 platforms.

On creation of a token, you're shown a URL which a platform can use to
initiate dynamic registration.
Dynamic registration without a token is not possible.
A token can only be used once.
Each token has an "intended purpose" field, to help identify the purpose
of tokens that aren't used.