don't automatically create AGS line items
Previously, an LTI 1.3 assignment and grade service line item was
automatically created if it didn't already exist.

This changes the `get_lti_13_lineitem` method to not automatically
create a line item by default.

If a resource has "when to report scores back" set to something that
would cause scores to be reported automatically, and it doesn't have an
associated grades line item, then a warning message is shown on the
resource dashboard. You can click a button to create a line item, or say
that you don't want one (for example, for formative material).

If you said you didn't want one and change your mind, there's a button
at the bottom of the settings page to create a line item.

If you click the button to report scores manually, a line item is
automatically created if necessary.