fix the blackboard registration process
When you register with the Anthology developer portal, it gives the
issuer 'https://developer.blackboard.com' in the dynamic registration
process.

However, launches from real Blackboard use 'https://blackboard.com' as
the issuer.

So the Blackboard registration form looks for a tool with the developer
address, and creates one for just blackboard.com with the same settings.