blackboard registration: copy all the fields from the dev tool
When you register with the Anthology developer portal, an LtiTool object is created with all the fields filled in.

The form which creates an LtiTool for the real https://blackboard.com issuer was only copying the auth_login_url.

This commit changes it to use the original object, but reset its pk so a new object is saved, and all the fields are kept except for the ones that are explicitly changed.