LTIConsumer.contexts_grouped_by_period: rewrite the query for num_attempts
The `num_attempts` annotation on contexts was calculated with a custom
call to django.db.models.Func. It turns out that postgres doesn't like
this: it complains that a field is not used in an aggregation.

This commit changes the query to more directly count Attempt objects.
It has two separate subqueries for each of LTI 1.1 and 1.3, but
otherwise it's neat.