lti 1.3 launch: use the LIS claim for consumer_user_id
The platform can use the Learning Information Services claim
(https://www.imsglobal.org/spec/lti/v1p3/#lislti) to give extra
information about the person doing the launch.

Canvas uses this to give the real username for the user. The `sub` value
it uses in the launch OAuth is just a UUID.

This changes the ResourceLaunchView to use the `person_sourcedid` field
from the LIS claim if that claim is present and the field is non-empty,
and save that in the LTIUserData's consumer_user_id field.