the scorm CMI response gives the PK of the attempt
When running a broken attempt, a new attempt is created automatically.

This changes the response from the scorm_cmi view to include the PK of
the attempt, so that the client will use the right address to report
changed SCORM elements to.

The response also includes the PK of the broken attempt if there was
one. The client could use this to show a message explaining what's
happened (but doesn't yet)