Use the .exam parser from the compiler to read .exam files
The Exam.source method uses the NumbasObject parser from the Numbas
compiler to read the source.exam file from an exam package, so that its
migrations are applied.

The immediate motivation for this is to update old packages to use the
new feedback settings, so that the dashboard code doesn't have to know
about both versions.

The parser code is copied directly into numbas_lti/examparser. It can be
updated by running `make update_examparser`. The path to the numbas
compiler is given by the variable NUMBAS_COMPILER_PATH