Improve the access change edit form
It now uses the "big" class that the settings form uses, so the
fieldsets are laid out horizontally when possible.

This commit adds some JavaScript to make the search field for the
members list filter the list of members.

There's a little change to the form.big styling so a help block at the
start of a fieldset gets the right margin-top.