Improve time reporting for scores in LTI 1.3
The start and submission times are added to the reported grade. This
needed a change in pylti1p3.

There's a new setting, REPORT_SCORE_SUBTRACT_MINUTES, which is
subtracted from the timestamp reported to the consumer.

If the attempt hasn't ended, the platform should use the timestamp
instead of the submittedAt time, so I think it's right to subtract from
them both.

I'm not sure if we should be using now for the timestamp, or the time of
the last SCORM element. The spec says that the timestamp field is meant
to prevent out-of-order reports from overwriting newer info, so I think
that using the time of the most recent SCORM element is correct.

If an attempt doesn't have any SCORM elements, the start_time is used
instead.